# device-app/api.yaml
openapi: '3.0.0'
info:
  title: Devices API
  version: '1.0.0'
paths:
  /devices/{id}:
    get:
      summary: Get device data by Id.
      parameters:
        - name: id
          in: path
          description: Device Id
          required: true
          schema:
            type: integer
            format: int64
          example:
            id: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  houseId:
                    type: integer
                  houseAddress:
                    type: string
                  deviceType:
                    type: string
                  serialNumber:
                    type: string
                  status:
                    type: string
                example:
                  id: 1
                  houseId: 10
                  houseAddress: "some city, some house"
                  deviceType: thermometer
                  serialNumber: cvb01kh3253
                  status: on
        '404':
          description: Device not found response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
        '500':
          description: Server error response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
  /devices/{id}/status:
    put:
      summary: Update device data
      parameters:
        - name: id
          in: path
          description: Device Id
          required: true
          schema:
            type: integer
          example:
            id: 1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              example:
                status: on
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  houseId:
                    type: integer
                  houseAddress:
                    type: string
                  deviceType:
                    type: string
                  serialNumber:
                    type: string
                  status:
                    type: string
                example:
                  id: 1
                  houseId: 10
                  houseAddress: "some city, some house"
                  deviceType: thermometer
                  serialNumber: cvb01kh3253
                  status: on
        '404':
          description: Device not found response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
        '500':
          description: Server error response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
  /devices/{id}/commands:
    post:
      summary: Push command to device
      parameters:
        - name: id
          in: path
          description: Device Id
          required: true
          schema:
            type: integer
          example:
            id: 1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                commandType:
                  type: string
                value:
                  type: integer
              example:
                commandType: increase
                value: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                example:
                  code: 200
                  message: 'OK'
        '404':
          description: Device not found response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
        '500':
          description: Server error response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string  